requirements.txt
asgiref==3.8.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==4.1.0
djangorestframework-simplejwt[token-blacklist]>=5.3.1
freezegun==1.4.0
Pillow==10.2.0
pytest-xdist==3.5.0